Syntx highlighting readability increased
